+++ madblog/markdown/_processors/__init__.py
@@ -1,7 +1,5 @@
 from .activitypub import MarkdownActivityPubMentions
 from .autolink import MarkdownAutolink
-from .latex import MarkdownLatex
-from .mermaid import MarkdownMermaid
 from .tags import MarkdownTags
 from .tasklist import MarkdownTaskList
 from .toc import MarkdownTocMarkers
@@ -10,8 +8,6 @@ from .toc import MarkdownTocMarkers
 __all__ = [
     "MarkdownActivityPubMentions",
     "MarkdownAutolink",
-    "MarkdownLatex",
-    "MarkdownMermaid",
     "MarkdownTaskList",
     "MarkdownTags",
     "MarkdownTocMarkers",